+++ broker/src/asapo_broker/database/mock_database.go
@@ -24,6 +24,11 @@ func (db *MockedDatabase) Ping() error {
 	return args.Error(0)
 }
 
+func (db *MockedDatabase) SetSettings(settings DBSettings) {
+	db.Called()
+}
+
+
 func (db *MockedDatabase) ProcessRequest(db_name string, data_collection_name string, group_id string, op string, extra_param string) (answer []byte, err error) {
 	args := db.Called(db_name, data_collection_name, group_id, op, extra_param)
 	return args.Get(0).([]byte), args.Error(1)