+++ deploy/asapo_orchestration_docker/scripts/resources.tf
@@ -4,4 +4,5 @@ resource "nomad_job" "monitoring" {
 
 resource "nomad_job" "nginx" {
   jobspec = "${data.template_file.nginx_template.rendered}"
+  depends_on = [nomad_job.monitoring,null_resource.influxdb,null_resource.grafana,null_resource.telegraf]
 }