+++ tests/manual/python_tests/consumer/consumer_api.py
@@ -7,14 +7,17 @@ source, path, beamtime, token = sys.argv[1:]
 broker = asapo_consumer.create_server_broker(source,path,False, beamtime,"",token,1000)
 group_id = broker.generate_group_id()
 
+res = broker.query_images("_id > 0", substream="1")
 
-_, meta = broker.get_by_id(1,group_id, meta_only=True)
+print(res)
+
+#data, meta = broker.get_by_id(5,group_id,"default", meta_only=False)
 
 #meta["buf_id"]=0
-data = broker.retrieve_data(meta)
+#data = broker.retrieve_data(meta)
 
-print (meta)
-print (len(data),data[0:100])
-data.tofile("out")
+#print (meta)
+#print (len(data),data[0:100])
+#data.tofile("out")
 
 sys.exit(0)